+++ common/src/main/java/net/mehvahdjukaar/every_compat/modules/create/CreateModule.java
@@ -1,4 +1,4 @@
-package net.mehvahdjukaar.every_compat.modules.fabric.create;
+package net.mehvahdjukaar.every_compat.modules.create;
 
 import com.simibubi.create.CreateClient;
 import com.simibubi.create.content.decoration.palettes.ConnectedGlassPaneBlock;
@@ -10,25 +10,31 @@ import net.mehvahdjukaar.every_compat.EveryCompat;
 import net.mehvahdjukaar.every_compat.api.RenderLayer;
 import net.mehvahdjukaar.every_compat.api.SimpleEntrySet;
 import net.mehvahdjukaar.every_compat.api.SimpleModule;
+import net.mehvahdjukaar.every_compat.dynamicpack.ServerDynamicResourcesHandler;
+import net.mehvahdjukaar.moonlight.api.platform.PlatHelper;
+import net.mehvahdjukaar.moonlight.api.resources.ResType;
 import net.mehvahdjukaar.moonlight.api.set.wood.WoodType;
 import net.mehvahdjukaar.moonlight.api.set.wood.WoodTypeRegistry;
 import net.mehvahdjukaar.moonlight.api.util.Utils;
 import net.minecraft.core.registries.Registries;
 import net.minecraft.resources.ResourceLocation;
+import net.minecraft.server.packs.resources.ResourceManager;
 import net.minecraft.tags.BlockTags;
+import net.minecraft.world.item.Item;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.Blocks;
 
-// SUPPORT: v0.5.1f +
+import java.util.Objects;
+
+// SUPPORT: v0.5.1+
 public class CreateModule extends SimpleModule {
 
     public final SimpleEntrySet<WoodType, Block> windows;
     public final SimpleEntrySet<WoodType, Block> windowPanes;
 
-
     public CreateModule(String modId) {
         super(modId, "c");
-        ResourceLocation tab = modRes("palettes");
+        var tab = modRes("palettes");
 
         windows = SimpleEntrySet.builder(WoodType.class, "window",
                         getModBlock("oak_window"), () -> WoodTypeRegistry.OAK_TYPE, //AllPaletteBlocks.OAK_WINDOW
@@ -38,7 +44,6 @@ public class CreateModule extends SimpleModule {
                 .defaultRecipe()
                 .setRenderType(RenderLayer.CUTOUT_MIPPED)
                 .createPaletteFromOak(p -> p.remove(p.getDarkest()))
-                //TEXTURES: planks
                 .addTextureM(modRes("block/palettes/oak_window"), EveryCompat.res("block/palettes/oak_window_m"))
                 .addTextureM(modRes("block/palettes/oak_window_connected"), EveryCompat.res("block/palettes/oak_window_connected_m"))
                 .build();
@@ -71,8 +76,7 @@ public class CreateModule extends SimpleModule {
     }
 
     @Environment(EnvType.CLIENT)
-    private static final class CreateClientModule {
-
+    private static class CreateClientModule {
         private static void clientStuff(CreateModule module) {
             module.windows.blocks.forEach((w, b) -> {
                 String path = "block/" + module.shortenedId() + "/" + w.getNamespace() + "/palettes/" + w.getTypeName() + "_window";
@@ -88,4 +92,53 @@ public class CreateModule extends SimpleModule {
         }
     }
 
+    @Override
+    // Recipes
+    public void addDynamicServerResources(ServerDynamicResourcesHandler handler, ResourceManager manager) {
+        super.addDynamicServerResources(handler, manager);
+        if (!PlatHelper.isModLoaded("sawmill")) {
+            for (WoodType w : WoodTypeRegistry.getTypes()) {
+                if (w.getBlockOfThis("slab") != null)
+                    sawRecipe(2, w.planks.asItem(), Objects.requireNonNull(w.getBlockOfThis("slab")).asItem(), null, w, handler);
+
+                if (w.getBlockOfThis("stairs") != null)
+                    sawRecipe(1, w.planks.asItem(), Objects.requireNonNull(w.getBlockOfThis("stairs")).asItem(), null, w, handler);
+
+                sawRecipe(6, w.planks.asItem(), null, new ResourceLocation("minecraft", "stick"), w, handler);
+            }
+        }
+
+    }
+
+    public void sawRecipe(int amount, Item input, Item output, ResourceLocation item, WoodType woodType, ServerDynamicResourcesHandler handler) {
+        String blank = """
+                {
+                    "type": "minecraft:stonecutting",
+                    "count": [amount],
+                    "ingredient": {
+                        "item": "[input]"
+                    },
+                    "result": "[output]"
+                }
+                """;
+
+        ResourceLocation resloc;
+        String recipe;
+        if (output != null) {
+            recipe = blank.replace("[amount]", String.valueOf(amount))
+                    .replace("[input]", Utils.getID(input).toString())
+                    .replace("[output]", Utils.getID(output).toString());
+            resloc = EveryCompat.res(
+                    shortenedId() + "/" + woodType.getNamespace() + "/" + output + "_from_" + input + "_stonecutting");
+        } else { // item != null
+            recipe = blank.replace("[amount]", String.valueOf(amount))
+                    .replace("[input]", Utils.getID(input).toString())
+                    .replace("[output]", item.toString());
+            resloc = EveryCompat.res(
+                    shortenedId() + "/" + woodType.getNamespace() + "/" + item.getPath() + "_from_" + input + "_stonecutting");
+        }
+
+        handler.dynamicPack.addBytes(resloc, recipe.getBytes(), ResType.RECIPES);
+
+    }
 }