+++ changelog.md
@@ -2,6 +2,7 @@
 - **More Crafting Table (LieOnLion)** (FABRIC): Updated to support v1.2.7
 - **Create** (COMMON): Fixed the crash because a method in CreateModule should be only on CLIENT side
 - **LANG** (COMMON): zn_ch - @ChuijkYahus
+- **Building But Better** (FORGE): **EveryComp** can work with either v1.0.1 or v1.1.1 of **Building But Better**
 
 ---